Move common fucntionality out of native console.
