+++ Code/Tools/TestImpactFramework/Frontend/Console/Common/Code/testimpactframework_frontend_console_common_headers_files.cmake
@@ -7,5 +7,5 @@
 #
 
 set(FILES
-    Source/TestImpactConsole.cpp
+    Include/Headers/TestImpactFramework/TestImpactConsoleMain.h
 )
