+++ Code/Tools/TestImpactFramework/Frontend/Console/Native/Code/CMakeLists.txt
@@ -24,8 +24,11 @@ ly_add_target(
         PRIVATE
             ${LY_TEST_IMPACT_CONFIG_FILE_PATH_DEFINITION}
     BUILD_DEPENDENCIES
-        PUBLIC
+        PRIVATE
+            AZ::TestImpact.Frontend.Console.Common.Static
             AZ::TestImpact.Runtime.Native.Static
+        PUBLIC
+            AZ::TestImpact.Frontend.Console.Common.Headers
 )
 
 ly_add_target(
