+++ Code/Tools/TestImpactFramework/Frontend/Console/Native/Code/testimpactframework_frontend_console_native_files.cmake
@@ -7,5 +7,4 @@
 #
 
 set(FILES
-    Source/TestImpactConsole.cpp
 )
