+++ Code/Tools/TestImpactFramework/Frontend/Console/Native/Code/testimpactframework_frontend_console_native_static_files.cmake
@@ -7,7 +7,6 @@
 #
 
 set(FILES
-    Include/TestImpactFramework/TestImpactConsoleMain.h
     Source/TestImpactCommandLineOptions.h
     Source/TestImpactCommandLineOptions.cpp
     Source/TestImpactCommandLineOptionsUtils.cpp
