+++ Code/Tools/TestImpactFramework/Frontend/Console/Native/Code/testimpactframework_frontend_console_native_static_tests_files.cmake
@@ -7,4 +7,9 @@
 #
 
 set(FILES
+    Tests/TestImpactCommandLineOptionsTest.cpp
+    Tests/TestImpactRuntimeConfigurationFactoryTest.cpp
+    Tests/TestImpactFrontendConsoleStaticTestMain.cpp
+    Tests/TestImpactConsoleMainTest.cpp
+    Tests/TestImpactConsoleTestSequenceEventHandlerTest.cpp
 )
